requirements.txt
flask>=2.3.0
opencv-python>=4.8.0
torch>=2.0.0
numpy>=1.24.0
werkzeug>=2.3.0
transformers>=4.48.0
accelerate>=0.26.0
pillow
timm
ftfy
regex
iopath
huggingface_hub